ho ho ho merry christmas  - removed a bunch of old and unused stuff    - LuaSetup, MessageBus, DopClasses, DopFunctions, PoolTools  - added enum helpers for Keyboard_Key and Mouse_Button  - got rid of Mouse_Wheel enum, event data now has x and y  - added enum helpers for Gamepad_Button and Gamepad_Axis  - added extra buttons an axes to Gamepad, it now exactly matches glfw  - got rid of all the macro crap in GuiWidgets, added a few minor things      - rearranged paramaters to custom widgets to match standard equivilants  - added is_gamepad_connected method to InputDevices  - allow Scene to skip updates for one frame after a long loading time  - cleaned up EnumHelper a bit  - increaded maximum argument count for SQEE_FOR_EACH from 50 to 100  - changed size of SmallString from 31 chars to 23  - added to_c_string overload for StackString  - optimised StackString even more because I can't control myself  - cleaned up StackVector and added some missing features  - added format_append helper function for fmt::format_to + std::back_inserter  - added SQEE_UNREACHABLE() macro to debug/Assert.hpp  - cleaned up traits code for Vector/Matrix/Quaternion  - fixed slerp function fallback to lerp, which had been broken forever  - added angular_difference, lerp_shortest, and lerp_guided quaternion functions    - lerp_guided is a hacky and dumb thing I needed for STS    - may remove it in the future if I come up with something better  - replaced SQEE_ENUM_JSON_CONVERSIONS macro with adl_serializer specialisation  - update nlohmann json to latest version    - disabled implcit conversions, they were causing issues    - some code is a bit uglier now but there's not much I can do  - modified wren to allow newlines before else    - sadly getting this change upstreamed was rejected for now  - added extra mouse buttons and horizontal scroll to imgui backend  - many, many improvements to WrenPlus:     - added FieldHelper template to easy binding member variables     - added macros for defining wren::Traits     - added macros for more readable binding code        - WrenPlusVM::register_method is still available for advanced use     - added methods to automatically generate and register pointer (in)equality methods     - foreign objects now store a type tag along with the pointer     - foreign class handles are now cached in a simple vector     - you can (and must) now specify the list of directories to search for modules in     - added SlotHelper for std::vector (get only for now)     - added SlotHelper for SQEE_ENUM_HELPER (as strings)     - added SlotHelper for sq::StackVector (get only for now)     - added set support for std::optional     - misc cleanup and optimisation

+++ include/sqee/objects/Armature.hpp
@@ -35,7 +35,6 @@ public: //====================================================//
     /// A set of tracks that can be sampled to get a Pose.
     struct SQEE_API Animation
     {
-        // todo: investigate using std::pmr for allocating tracks
         struct Track
         {
             TinyString key;
@@ -57,7 +56,7 @@ public: //====================================================//
 
     //--------------------------------------------------------//
 
-    uint get_bone_count() const { return uint(mRestPose.size()); }
+    size_t get_bone_count() const { return mRestPose.size(); }
 
     const Pose& get_rest_pose() const { return mRestPose; }
 
@@ -82,6 +81,8 @@ public: //====================================================//
 
     //--------------------------------------------------------//
 
+    // todo: poses are big objects and should probably not be reallocated constantly
+
     static Pose blend_poses(const Pose& a, const Pose& b, float factor);
 
     static Pose compute_pose_continuous(const Animation& animation, float time);
@@ -101,7 +102,7 @@ public: //====================================================//
     std::vector<Mat4F> compute_skeleton_matrices(const Pose& pose) const;
 
     /// Compute transposed pose matrices, ready for a shader.
-    void compute_ubo_data(const Pose& pose, Mat34F* out, uint len) const;
+    void compute_ubo_data(const Pose& pose, Mat34F* out, size_t len) const;
 
 private: //===================================================//
